+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed: Improved workflow compatibility with ACF Extended frontend forms by triggering only when the required metadata has already been processed (Issue #1649).
 - Fixed: Fatal error on Scheduled Actions screen when Action Scheduler 4.1 is loaded (Issue #1684).
+- Fixed: Activating a Workflow Sets Status to "Scheduled" Instead of "Active" on Sites with a Positive UTC Timezone Offset (Issue #1698).
 - Added: Copy workflow to text feature for development, support or debug work.
 
 [4.10.4] - 11 Aug, 2026